# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Music API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/telegram": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Telegram",
        "operationId": "login_telegram_v1_auth_telegram_post",
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramLoginIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Token",
        "operationId": "refresh_token_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "User-Agent",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Agent"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me/lang": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Set Lang",
        "operationId": "set_lang_v1_me_lang_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LangIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/admin/crawler/channels": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Crawler Channels",
        "operationId": "crawler_channels_admin_crawler_channels_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlChannelOut"
                  },
                  "title": "Response Crawler Channels Admin Crawler Channels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/library/channels": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "My Channels",
        "operationId": "my_channels_v1_library_channels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserChannelOut"
                  },
                  "type": "array",
                  "title": "Response My Channels V1 Library Channels Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Add Channel",
        "operationId": "add_channel_v1_library_channels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChannelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddChannelOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/channels/suggest": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Suggest Channel",
        "description": "Ask for a channel we do not index yet (ADR-002 §3, the \"user\" source).\n\nA channel somebody already indexed is not a request at all: it is attached to the\nasker's library immediately, because channels are indexed once for the platform.",
        "operationId": "suggest_channel_v1_channels_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChannelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestChannelOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/library/channels/{channel_id}": {
      "delete": {
        "tags": [
          "channels"
        ],
        "summary": "Remove Channel",
        "operationId": "remove_channel_v1_library_channels__channel_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/channels/categories": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Categories",
        "operationId": "categories_v1_channels_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "type": "array",
                  "title": "Response Categories V1 Channels Categories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/channels/featured": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Featured",
        "operationId": "featured_v1_channels_featured_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ChannelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/channels/{channel_id}": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Channel Detail",
        "operationId": "channel_detail_v1_channels__channel_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddChannelIn": {
        "properties": {
          "ref": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Ref",
            "description": "@username or t.me link"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ref"
        ],
        "title": "AddChannelIn"
      },
      "AddChannelOut": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/ChannelOut"
          },
          "created": {
            "type": "boolean",
            "title": "Created"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "created"
        ],
        "title": "AddChannelOut"
      },
...
      "CategoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name_fa": {
            "type": "string",
            "title": "Name Fa"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "slug",
          "name_fa",
          "name_en"
        ],
        "title": "CategoryOut"
      },
...
      "CrawlChannelOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "crawl_status": {
            "type": "string",
            "title": "Crawl Status"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "oldest_crawled_msg_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Crawled Msg Id"
          },
          "newest_crawled_msg_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newest Crawled Msg Id"
          },
          "tracks_count": {
            "type": "integer",
            "title": "Tracks Count"
          },
          "last_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Crawl At"
          },
          "next_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Crawl At"
          },
          "crawl_interval_sec": {
            "type": "integer",
            "title": "Crawl Interval Sec"
          },
          "fail_count": {
            "type": "integer",
            "title": "Fail Count"
          },
          "crawl_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crawl Error"
          },
          "extraction_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Rate"
          },
          "preview_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "crawl_status",
          "progress_pct",
          "tracks_count",
          "crawl_interval_sec",
          "fail_count"
        ],
        "title": "CrawlChannelOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LangIn": {
        "properties": {
          "lang": {
            "type": "string",
            "enum": [
              "fa",
              "en"
            ],
            "title": "Lang"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lang"
        ],
        "title": "LangIn"
      },
...
      "MeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tg_id": {
            "type": "integer",
            "title": "Tg Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "lang": {
            "type": "string",
            "enum": [
              "fa",
              "en"
            ],
            "title": "Lang"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "premium_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Premium Until"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits"
          },
          "referral_code": {
            "type": "string",
            "title": "Referral Code"
          },
          "public_profile": {
            "type": "boolean",
            "title": "Public Profile",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tg_id",
          "username",
          "first_name",
          "lang",
          "plan",
          "premium_until",
          "features",
          "limits",
          "referral_code"
        ],
        "title": "MeOut"
      },
...
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn"
      },
...
      "SuggestChannelOut": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "subscribed",
              "queued"
            ],
            "title": "Status"
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChannelOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SuggestChannelOut",
        "description": "Either we already index it (and you now have it), or your request is queued."
      },
...
      "TelegramLoginIn": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 10,
            "title": "Init Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "TelegramLoginIn"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "access_expires_at": {
            "type": "integer",
            "title": "Access Expires At"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "refresh_expires_at": {
            "type": "integer",
            "title": "Refresh Expires At"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "start_param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Param"
          },
          "me": {
            "$ref": "#/components/schemas/MeOut"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "access_expires_at",
          "refresh_token",
          "refresh_expires_at",
          "me"
        ],
        "title": "TokenOut"
      },
...
      "UserChannelOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "indexing",
              "active",
              "paused",
              "failed",
              "blacklisted"
            ],
            "title": "Status"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "tracks_count": {
            "type": "integer",
            "title": "Tracks Count"
          },
          "subscribers_count": {
            "type": "integer",
            "title": "Subscribers Count"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "mtproto",
              "bot_admin"
            ],
            "title": "Source"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "title",
          "status",
          "status_reason",
          "progress_pct",
          "tracks_count",
          "subscribers_count",
          "is_featured",
          "category_id",
          "source",
          "added_at"
        ],
        "title": "UserChannelOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}